ni-119 improve multipattern and add path copying

+++ src/sockets/editor/handlers.ts
@@ -2,7 +2,6 @@ import { Server, Socket } from "socket.io";
 import { type Processes } from "../../lib/types";
 import { makeChanges } from "../../lib/make-changes";
 import { Configuration } from "../../lib/configuration";
-import path from "path";
 
 const processes: Processes = {};
 
@@ -10,11 +9,8 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
     const edit = (data: { key: string; value: string; fileKey: string; path: string }) => {
         const key = data.key;
         const value = data.value;
-        const fileKey = data.fileKey;
         const filePath = data.path;
-        const fileData = config.files.find(
-            (f) => f.path === path.join(process.cwd(), filePath).replaceAll(path.sep, "/"),
-        );
+        const fileData = config.files.find((f) => f.path === filePath);
 
         if (!fileData) return;
 
@@ -26,7 +22,6 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
                 value: value || "",
             },
             {
-                fileKey,
                 filePath: fileData.path,
                 indent: config.indent,
             },
@@ -34,7 +29,7 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
         );
     };
 
-    const create = (data: { key: string; value: string; fileKey: string }) => {
+    const create = (data: { key: string }) => {
         const key = data.key;
         console.log(`Create key "${key}"`);
         config.files.forEach((fileData) => {
@@ -44,7 +39,6 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
                     type: "create",
                 },
                 {
-                    fileKey: fileData.key,
                     filePath: fileData.path,
                     indent: config.indent,
                 },
@@ -53,22 +47,24 @@ export const editorHandlers = (_io: Server, socket: Socket, config: Configuratio
         });
     };
 
-    const remove = (data: { key: string; value: string; fileKey: string }) => {
+    const remove = (data: { key: string; staticPart: string }) => {
         const key = data.key;
+        const staticPart = data.staticPart;
         console.log(`Delete key "${key}"`);
         config.files.forEach((fileData) => {
-            makeChanges(
-                {
-                    key,
-                    type: "delete",
-                },
-                {
-                    fileKey: fileData.key,
-                    filePath: fileData.path,
-                    indent: config.indent,
-                },
-                processes,
-            );
+            if (fileData.staticPart === staticPart) {
+                makeChanges(
+                    {
+                        key,
+                        type: "delete",
+                    },
+                    {
+                        filePath: fileData.path,
+                        indent: config.indent,
+                    },
+                    processes,
+                );
+            }
         });
     };
